--- gradle.properties
@@ -1,1 +1,0 @@
-kotlin.code.style=official
