--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Wed Oct 29 15:02:52 CET 2025
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.14-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
