--- settings.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
-}
-rootProject.name = "ubelluris"
