+++ src/main/kotlin/org/entur/ror/ubelluris/filter/FilterConfig.kt
@@ -1,4 +1,4 @@
-package filter
+package org.entur.ror.ubelluris.filter
 
 import org.entur.netex.tools.lib.config.FilterConfig
