+++ src/main/kotlin/org/entur/ror/ubelluris/handlers/StopPlaceIdHandler.kt
@@ -1,13 +1,13 @@
-package plugin.handlers
+package org.entur.ror.ubelluris.handlers
 
-import model.NetexTypes
+import org.entur.ror.ubelluris.model.NetexTypes
 import org.entur.netex.tools.lib.extensions.addNewAttribute
 import org.entur.netex.tools.lib.output.DelegatingXMLElementWriter
 import org.entur.netex.tools.lib.output.XMLElementHandler
 import org.xml.sax.Attributes
 import org.xml.sax.helpers.AttributesImpl
 
-class StopPlaceParentSiteRefHandler(
+class StopPlaceIdHandler(
 ) : XMLElementHandler {
     override fun startElement(
         uri: String?,
@@ -16,11 +16,11 @@ class StopPlaceParentSiteRefHandler(
         attributes: Attributes?,
         writer: DelegatingXMLElementWriter
     ) {
-        val idValue = attributes?.getValue("ref")?.replace("SE:050", "SAM") ?: ""
+        val idValue = attributes?.getValue("id")?.replace("SE:050", "SAM") ?: ""
         val newAttributes = AttributesImpl()
-        newAttributes.addNewAttribute("ref", idValue)
+        newAttributes.addNewAttribute("id", idValue)
         newAttributes.addNewAttribute("version", "1")
-        writer.startElement(uri, NetexTypes.PARENT_SITE_REF, NetexTypes.PARENT_SITE_REF, newAttributes)
+        writer.startElement(uri, NetexTypes.STOP_PLACE, NetexTypes.STOP_PLACE, newAttributes)
     }
 
     override fun characters(
@@ -38,6 +38,6 @@ class StopPlaceParentSiteRefHandler(
         qName: String?,
         writer: DelegatingXMLElementWriter
     ) {
-        writer.endElement(uri, NetexTypes.PARENT_SITE_REF, NetexTypes.PARENT_SITE_REF)
+        writer.endElement(uri, NetexTypes.STOP_PLACE, NetexTypes.STOP_PLACE)
     }
 }
